Test behaviour against newly generated app
Adopt the approach (and much of the support code) taken by hanami in their CLI testing. Now, instead of having a "dummy" app baked into our spec/ dir, we generate a new dry-web-roda app completely from scratch, following the approach a user would take as closely as possible, including bundling all the gems for the app. With the app generated, our first test is to actually boot it up and then ensure a page renders as we expect. This should serve as a helpful smoke test for setup/configuration/dependency issues in the future.