Use container from config to fetch rack-monitor
We can't rely on the return value from `.configure` here, since it  returns the value of the last applied setting (which won't always be our container).

This fixes the bug that will arise if a Dry::Web::Roda::Application is configured and the container is not the last setting applied.