Rename Application classes to Web
One of the big ideas we’re promoting with this application structure is that the _web interface is not your app._ Having the web interface class named `Application` kind of interfered with that message. `Web` is a much clearer demonstration of its purpose: it’s just one of (potentially many) interfaces to your core app.