Change Database Cleaner strategy
I've an API project with 1700+ tests and it's growing. It used to take 15 to 20 minutes to run the complete suite. So I wanted to avoid running the suite at all cost. After a single change as suggested in this PR, the time dropped to 40 seconds. There's still a `truncation` strategy in the before suite hook. But when the tests start to run, transaction strategy is good enough.