Track to be generated .env* files as regular blobs
At some point in the past, someone had to do something like this:

```bash
git add --force lib/dry/web/roda/templates/.env*
```

This is because in `lib/dry/web/roda/templates/` there was a
`.gitignore` file with the intention to be added to the generated
application. However, git (of course) was treating it as a regular
.gitignore file so it was ignoring everything it declared, which
included `.env*` files in that directory.

If in the future we happen to rename the directory
`lib/dry/web/roda/templates/` to anything else, `.env*` files would be
ignored and then not generated.

The solution is to just make .gitignore in that directory a template
file.